+++ src/bun.js/bindings/ZigGlobalObject.cpp
@@ -4072,21 +4072,6 @@ GlobalObject::PromiseFunctions GlobalObject::promiseHandlerID(Zig::FFIFunction h
     }
 }
 
-void* vm(Zig::GlobalObject* globalObject)
-{
-    return globalObject->bunVM();
-}
-
-void* vm(JSC::VM& vm)
-{
-    return WebCore::clientData(vm)->bunVM;
-}
-
-void* vm(JSC::JSGlobalObject* lexicalGlobalObject)
-{
-    return WebCore::clientData(lexicalGlobalObject->vm())->bunVM;
-}
-
 #include "ZigGeneratedClasses+lazyStructureImpl.h"
 #include "ZigGlobalObject.lut.h"
